extracted file operations in to common module
