+++ settings.gradle
@@ -26,9 +26,11 @@ include("submission:submission-security")
 include("submission:notifications")
 include("submission:stats")
 include("submission:submission-handlers")
-include("submission:persistence-common-api")
-include("submission:persistence")
 include("submission:submission-config")
+
+include("submission:persistence-common-api")
+include('submission:persistence-common')
+include("submission:persistence")
 
 include("scheduler:cluster-client")
 include("scheduler:pmc-processor-task")
