+++ submission/persistence/build.gradle.kts
@@ -28,8 +28,10 @@ dependencies {
     api(project(":commons:commons-model-extended"))
     api(project(":commons:commons-model-extended-mapping"))
     api(project(":commons:commons-test"))
-    api(project(":submission:persistence-common-api"))
     api(project(":submission:submission-config"))
+
+    api(project(":submission:persistence-common-api"))
+    api(project(":submission:persistence-common"))
 
     implementation(SpringAutoConfigure)
     implementation(SpringDataJpa)
